Create complier.py
Short and simple python program for the compiling of python scripts (.py) to the byte code compiled python (.pyc)

Python is an interpreted language, as opposed to a compiled one, though the distinction can be blurry because of the presence of the bytecode compiler. This means that source files can be run directly without explicitly creating an executable which is then run.